% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawRegionProportion_withProp.R
\name{DrawRegionProportion_withProp}
\alias{DrawRegionProportion_withProp}
\title{Draw regional cell type distribution with cellular proportion information}
\usage{
DrawRegionProportion_withProp(
  sce,
  label = "CARD_CellType",
  selCenter = seq_len(10)
)
}
\arguments{
\item{sce}{A single cell experiment object.}

\item{label}{A string character for the cell type variable.}

\item{selCenter}{A vector of the interested ROIs, e.g., 1:4.}
}
\value{
A plot object.
}
\description{
Draw regional cell type distribution with cellular proportion information
}
\examples{
data("example_sce")
DrawRegionProportion_withProp(example_sce,
                             label = "Proportions",
                             selCenter = 1:3)


}
