% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sce.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example single cell experiment for input}
\format{
A SingleCellExperiment object.
}
\usage{
data(example_sce)
}
\value{
A SingleCellExperiment object.
}
\description{
A simulated example input data file
}
\examples{
data(example_sce)
}
\keyword{datasets}
