% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRes.R
\docType{data}
\name{exampleRes}
\alias{exampleRes}
\title{Example DE output}
\format{
A list object.
}
\usage{
data(exampleRes)
}
\value{
A list object.
}
\description{
A simulated example DE output file
}
\examples{
data(exampleRes)
}
\keyword{datasets}
