% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportions.R
\name{getProportion}
\alias{getProportion}
\title{Define an accessor method for Proportion_CARD}
\usage{
getProportion(card)
}
\arguments{
\item{card}{A CARD object.}
}
\value{
A matrix containing the spot-level cell type proportion information
}
\description{
Define an accessor method for Proportion_CARD
}
\examples{
# getProportion(card)

}
