% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualSelectCenter.R
\name{ManualSelectCenter}
\alias{ManualSelectCenter}
\title{Manually select top ROIs}
\usage{
ManualSelectCenter(sce)
}
\arguments{
\item{sce}{A single cell experiment object.}
}
\value{
An sce object with selected centers and radiuses.
}
\description{
Manually select top ROIs
}
\examples{
data("example_sce")
example_sce <- BayesSpace::spatialPreprocess(example_sce, platform="Visium", log.normalize=TRUE)
# I commented this out because the shiny app will get stuck without input.
# example_sce <- ManualSelectCenter(example_sce)

}
