% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_hallmark.R
\docType{data}
\name{pathways_hallmark}
\alias{pathways_hallmark}
\title{Hallmark database}
\format{
A list object.
}
\source{
\href{http://www.gsea-msigdb.org/gsea/msigdb/annotate.jsp}{MSigDB}
}
\usage{
data(pathways_hallmark)
}
\description{
Hallmark database downloaded from MSigDB (Feb, 2023)
}
\examples{
data(pathways_hallmark)
}
\references{
Liberzon et al. (2015) Cell Syst. 1(6):417-425
(\href{https://pubmed.ncbi.nlm.nih.gov/26771021/}{PubMed})
}
\keyword{datasets}
