% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_kegg.R
\docType{data}
\name{pathways_kegg}
\alias{pathways_kegg}
\title{KEGG database}
\format{
A list object.
}
\source{
\href{http://www.gsea-msigdb.org/gsea/msigdb/annotate.jsp}{MSigDB}
}
\usage{
data(pathways_kegg)
}
\description{
KEGG database downloaded from MSigDB (Feb, 2023)
}
\examples{
data(pathways_kegg)
}
\references{
Kanehisa and Goto (2000) Nucleic Acids Research 28(1):27-30
(\href{https://pubmed.ncbi.nlm.nih.gov/10592173/}{PubMed})
}
\keyword{datasets}
