% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways_reactome.R
\docType{data}
\name{pathways_reactome}
\alias{pathways_reactome}
\title{REACTOME database}
\format{
A list object.
}
\source{
\href{http://www.gsea-msigdb.org/gsea/msigdb/annotate.jsp}{MSigDB}
}
\usage{
data(pathways_reactome)
}
\value{
A list object.
}
\description{
REACTOME database downloaded from MSigDB (Feb, 2023)
}
\examples{
data(pathways_reactome)
}
\references{
Jassal et al. (2020) Nucleic Acids Research 28(1):27-30
(\href{https://pubmed.ncbi.nlm.nih.gov/31691815/}{PubMed})
}
\keyword{datasets}
