% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOneSelectCenter.R
\name{PlotOneSelectedCenter}
\alias{PlotOneSelectedCenter}
\title{Plot one selected ROI}
\usage{
PlotOneSelectedCenter(sce, ploti, enhanced = FALSE)
}
\arguments{
\item{sce}{A single cell experiment object.}

\item{ploti}{A number of indicate which ROI to plot.}

\item{enhanced}{A logical variable for using enhanced data or not.}
}
\value{
A figure object for the selected ROI.
}
\description{
Plot one selected ROI
}
\examples{
data("example_sce")
example_sce <- mySpatialPreprocess(example_sce, platform="Visium")
example_sce$"array_col" <- example_sce$col
example_sce$"array_row" <- example_sce$row
example_sce$"pxl_col_in_fullres" <- example_sce$imagecol
example_sce$"pxl_row_in_fullres" <- example_sce$imagerow
PlotOneSelectedCenter(example_sce, ploti = 1)

}
